requirements.txt

fastapi==0.112.2
uvicorn==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
openai==1.43.0
numpy==1.26.4
faiss-cpu==1.8.0.post1
pypdf==4.3.1
httpx==0.27.0

